Move to realms/:realm_id/boards
