+++ server/realms/sql/index.ts
@@ -6,7 +6,7 @@ const dismissNotifications = `
         (SELECT id FROM users WHERE users.firebase_id = $/firebase_id/),
         (SELECT id FROM realms WHERE realms.string_id = $/realm_id/),
          DEFAULT)
-    ON CONFLICT(user_id, realm_id) DO UPDATE 
+    ON CONFLICT(user_id, realm_id) DO UPDATE
         SET dismiss_request_time = DEFAULT
         WHERE dismiss_notifications_requests.user_id = (SELECT id FROM users WHERE users.firebase_id = $/firebase_id/)
           AND dismiss_notifications_requests.realm_id = (SELECT id FROM realms WHERE realms.string_id = $/realm_id/)`;
@@ -15,11 +15,11 @@ const getRealmByExternalId = `
     SELECT * FROM realms WHERE string_id = $/realm_id/`;
 
 const getUserPermissionsForRealm = `
-    SELECT to_json(permissions) AS permissions 
+    SELECT to_json(permissions) AS permissions
     FROM users
       JOIN realm_user_roles ON users.id = realm_user_roles.user_id
       JOIN roles ON realm_user_roles.role_id = roles.id
-      JOIN realms ON realm_user_roles.realm_id = realms.id 
+      JOIN realms ON realm_user_roles.realm_id = realms.id
     WHERE users.firebase_id = $/user_id/ AND realms.string_id = $/realm_external_id/`;
 
 const createRealmInvite = `
@@ -33,27 +33,27 @@ const createRealmInvite = `
       INTERVAL '1 WEEK')`;
 
 const getInviteDetails = `
-    SELECT 
+    SELECT
       realms.string_id AS realm_id,
       inviter,
       invitee_email,
       created + duration < NOW() as expired,
-      used 
+      used
     FROM account_invites
-    JOIN realms ON account_invites.realm_id = realms.id 
-    WHERE nonce = $/nonce/ 
+    JOIN realms ON account_invites.realm_id = realms.id
+    WHERE nonce = $/nonce/
     ORDER BY created LIMIT 1`;
 
 const getInvites = `
-    SELECT 
+    SELECT
       nonce,
       users.firebase_id AS inviter_id,
       invitee_email,
       created + duration AS expires_at,
       created AS created_at,
-      label 
+      label
     FROM account_invites
-    JOIN realms ON account_invites.realm_id = realms.id 
+    JOIN realms ON account_invites.realm_id = realms.id
     JOIN users ON account_invites.inviter = users.id
     WHERE realms.string_id = $/realmExternalId/ AND used = false AND created + duration > NOW()`;
 
@@ -84,6 +84,19 @@ const fetchRolesInRealm = `
   INNER JOIN users ON users.id=realm_user_roles.user_id
   WHERE realms.string_id = $/realm_external_id/`;
 
+const createBoard = `
+    INSERT INTO boards(string_id, parent_realm_id, board_category_id, slug, tagline, avatar_reference_id, settings)
+    VALUES (
+        $/string_id/,
+        (SELECT id FROM realms WHERE realms.string_id = $/realm_id/),
+        (SELECT id FROM board_categories WHERE board_categories.string_id = $/category_id /),
+        $/slug/,
+        $/tagline/,
+        $/avatar_reference_id/,
+        $/settings/
+    )
+    RETURNING string_id;`;
+
 export default {
   getRealmBySlug: new QueryFile(path.join(__dirname, "realm-by-slug.sql")),
   dismissNotifications,
@@ -95,4 +108,5 @@ export default {
   addUserToRealm,
   findUserOnRealm,
   fetchRolesInRealm,
+  createBoard,
 };
