+++ server/realms/tests/create-board.test.ts
@@ -10,18 +10,21 @@ import {
 } from "utils/test-utils";
 
 import { GenericResponse } from "types/rest/responses";
-import { mocked } from "ts-jest/utils";
+import { TWISTED_MINDS_REALM_EXTERNAL_ID } from "test/data/realms";
 import request from "supertest";
-import router from "../../routes";
+import router from "../routes";
 
 jest.mock("handlers/auth");
 jest.mock("server/cache");
 jest.mock("server/db-pool");
 jest.mock("axios");
 
+// Uses the format: /:realm_id/board
+const CREATE_BOARD_ROUTE = `/${TWISTED_MINDS_REALM_EXTERNAL_ID}/boards`;
+
 const CREATE_BOARD_REQUEST = {
   slug: "new-board",
-  category_id: 1,
+  category_id: "9449bcab-fa55-4b0e-9ce0-438501e5fa79",
   tagline: "tagline",
   avatar_url: "https://example.com/avatar.png",
   settings: '{ "accentColor": "#7b00ff"}',
@@ -33,7 +36,7 @@ describe("#POST /board", () => {
   test("fails when user is unauthenticated", async () => {
     await wrapWithTransaction(async () => {
       const res = await request(server.app)
-        .post("/")
+        .post(CREATE_BOARD_ROUTE)
         .send(CREATE_BOARD_REQUEST);
 
       expect(res.status).toBe(401);
@@ -41,22 +44,16 @@ describe("#POST /board", () => {
     });
   });
 
-  test.todo(
-    "fails when route does not have a realm id"
-    // async () => {
-    // await wrapWithTransaction(async () => {
-    //   setLoggedInUser(BOBATAN_USER_ID);
-    //   const res = await request(server.app)
-    //     .post("/")
-    //     .send(CREATE_BOARD_REQUEST);
+  test("fails when route does not have a valid realm id", async () => {
+    await wrapWithTransaction(async () => {
+      setLoggedInUser(BOBATAN_USER_ID);
+      const res = await request(server.app)
+        .post("/realms/invalid/boards")
+        .send(CREATE_BOARD_REQUEST);
 
-    //   expect(res.status).toBe(500);
-    //   expect(res.body).toEqual(
-    //     "Realm permissions can only be fetched on a route that includes a realm id."
-    //   );
-    // });
-    // }
-  );
+      expect(res.status).toBe(404);
+    });
+  });
 
   test.todo(
     "fails when user does not have permission to create board"
@@ -64,7 +61,7 @@ describe("#POST /board", () => {
     //   await wrapWithTransaction(async () => {
     //     setLoggedInUser(ANON_WITH_NO_NAME_USER_ID);
     //     const res = await request(server.app)
-    //       .post("/")
+    //       .post(CREATE_BOARD_ROUTE)
     //       .send(CREATE_BOARD_REQUEST);
 
     //     expect(res.status).toBe(403);
@@ -79,7 +76,7 @@ describe("#POST /board", () => {
     await wrapWithTransaction(async () => {
       setLoggedInUser(BOBATAN_USER_ID);
       const res = await request(server.app)
-        .post("/")
+        .post(CREATE_BOARD_ROUTE)
         .send(CREATE_BOARD_REQUEST);
 
       expect(res.status).toBe(201);
